decoupled some code to make it easier to parse DVD chapters (although it's not done)

+++ script.js
@@ -12,6 +12,7 @@ var config = require('config'),
     fs = require('fs'),
     _ = require('lodash'),
     // ProgressBar = require('ascii-progress'),
+    _cliProgress = require('cli-progress'),
     Set = require('collections/set'),
     async = require('async'),
     inputFileName = checkInputFileName(argv.input || argv.disc || argv.bluray || argv.iso || argv.inputFileName),
@@ -24,20 +25,26 @@ var config = require('config'),
     dotMD5FileLocation = DISC_IMAGES_DIRECTORY + dotMD5FileName,
     outputFileLocation = '',
     baseHandbrakeParams = {
-        encoder : 'x265',
-        quality : '24.0',
-        aencoder : 'av_aac',
-        'custom-anamorphic' : true,
-        'keep-display-aspect' : true,
+        // encoder : 'x265',
+        // quality : '24.0',
+        // aencoder : 'av_aac',
+        // 'custom-anamorphic' : true,
+        // 'keep-display-aspect' : true,
+        // 'preset' : 'Roku 2160p60 4K HEVC Surround',
+        'preset' : 'Apple 2160p60 4K HEVC Surround',
+        'audio' : '1',
+        //'chapters' : '1-3',
+        //'preset' : 'Fast 1080p30',
         'subtitle' : 'scan',
         'subtitle-forced' : '1',
         'subtitle-burn' : '1',
-        'no-usage-stats' : '1',
-        rate : '29.97'
-    };
+        //'no-usage-stats' : '1',
+        // rate : '29.97'
+    },
+    TV_SHOW_MAX_DURATION = 60 * 60;
 
 function checkInputFileName(f) {
-    console.log(DISC_IMAGES_DIRECTORY);
+    //console.log(DISC_IMAGES_DIRECTORY);
     if (!f) {
         throw new Error('must pass --input or --disc or --bluray or --iso or --inputFileName');
     }
@@ -66,6 +73,9 @@ if (fs.existsSync(outputFileLocation)) {
     return;
 }
 
+function getChapterNumberToDuration(output) {
+}
+
 function getChapters(inputLocation, outputLocation) {
     return new Promise(function(resolve, reject) {
         var hbjs = require('handbrake-js'),
@@ -73,35 +83,84 @@ function getChapters(inputLocation, outputLocation) {
                 input : inputLocation,
                 output : outputLocation,
                 'min-duration' : (60 * 18),
+		        //'max-duration' : (60 * 55),
                 'no-usage-stats' : '1',
                 t : 0
             },
+            chapterNumberToDuration = {},
             chapters = [];
-    
-        hbjs._usage.disable();
 
+        //hbjs._usage.disable();
+
+        /*
         console.log('input path: ' + inputLocation);
         console.log('output path: ' + outputLocation);
+        console.log('mediaType: ' + mediaType);
+        */
 
         if (mediaType !== 'tv') {
             resolve(chapters);
             return;
         }
 
-        hbjs.spawn(params).on('error', function (err) {
-            console.log('error: ' + err);
-        }).on('output', function (output) {
-            _.each(output.split('\n'), function (line) {
-                var matchingChapter;
-                if (line.indexOf('+ title') === 0) {
-                    matchingChapter = line.match(/\+ title (\d+)/)[1];
-                    console.log(matchingChapter);
-                    chapters.push(matchingChapter);
-                }
+        delete params['no-usage-stats']
+        console.log('params: ' + JSON.stringify(params));
+
+        if (!argv.bluray) {
+            hbjs.exec(params, function (err, output, stderr) {
+                console.log('output: ' + output);
+                console.log('stderr: ' + stderr);
             });
-        }).on('complete', function () {
-            resolve(chapters);
-        });
+        } else {
+            hbjs.spawn(params).on('error', function (err) {
+                console.log('error: ' + err);
+            }).on('output', function (output) {
+                var matchingChapter, durationHours, durationMinutes, durationSeconds, totalDuration;
+                console.log('output: ' + output);
+                _.each(output.split('\n'), function (line) {
+                    if (line.indexOf('+ title') === 0) {
+                        matchingChapter = line.match(/\+ title (\d+)/)[1];
+                        /*
+                        console.log('matchingChapter: ' + matchingChapter);
+                        */
+                    }
+                    if (line.indexOf('  + duration') === 0) {
+                        /*
+                        console.log('line: ' + line);
+                        */
+                        durationHours = parseInt(line.match(/  \+ duration: 0?(\d+):0?(\d+):0?(\d+)/)[1]);
+                        durationMinutes = parseInt(line.match(/  \+ duration: 0?(\d+):0?(\d+):0?(\d+)/)[2]);
+                        durationSeconds = parseInt(line.match(/  \+ duration: 0?(\d+):0?(\d+):0?(\d+)/)[3]);
+                        /*
+                        console.log('durationHours: ' + durationHours);
+                        console.log('durationMinutes: ' + durationMinutes);
+                        console.log('durationSeconds: ' + durationSeconds);
+                        */
+
+                        totalDuration = ((60 * 60) * durationHours) + (60 * durationMinutes) + durationSeconds;
+                        /*
+                        console.log('matchingChapter: ' + matchingChapter);
+                        console.log('totalDuration: ' + totalDuration);
+                        */
+                        chapterNumberToDuration[matchingChapter] = totalDuration;
+                    }
+                });
+            }).on('complete', function () {
+                // console.log(chapterNumberToDuration);
+                _.each(_.keys(chapterNumberToDuration), function (chapterNumber) {
+                    // console.log(chapterNumber);
+                    if (chapterNumberToDuration[chapterNumber] < TV_SHOW_MAX_DURATION) {
+                        chapters.push(chapterNumber);
+                    }
+                });
+                /*
+                console.log(chapters);
+                */
+
+                // chapters.push(matchingChapter);
+                resolve(chapters);
+            });
+        }
     });
 }
 
@@ -109,13 +168,14 @@ function runHandbrake(inputLocation, outputLocation, chapters) {
     var hbjs = require('handbrake-js'),
         current = -1,
         fileParams = {
-            'no-usage-stats' : '1',
+            //'no-usage-stats' : '1',
             input : inputLocation,
             output : outputLocation,
+            verbose : 0
         },
         ripParams = _.extend({}, baseHandbrakeParams, fileParams),
         paramsList = [];
-        
+
     if (chapters.length === 0) {
         ripParams['main-feature'] = true;
         paramsList.push(ripParams);
@@ -123,33 +183,60 @@ function runHandbrake(inputLocation, outputLocation, chapters) {
         paramsList = _.map(chapters, function(chapterNumber, index) {
             return _.extend({}, ripParams, {
                 title : chapterNumber,
-                output : outputLocation.replace('DDD', _.padStart(index + 1, 3, '0')) 
+                output : outputLocation.replace('DDD', _.padStart(index + 1, 3, '0'))
             });
         });
     }
-    
-    hbjs._usage.disable();
-    async.eachLimit(paramsList, 1, function (params, nextRip) {
-        var percentComplete = 0;
-        console.log('params: ' + JSON.stringify(params));
 
+    // hbjs._usage.disable();
+    async.eachLimit(paramsList, 1, function (params, nextRip) {
+        var percentComplete, bar1;
+
+        // console.log('params: ' + JSON.stringify(params));
+	// nextRip(null, params);
         if (fs.existsSync(params.output)) {
             console.log('output already exists: ' + params.output);
             nextRip(null, params);
             return;
         }
 
+
         hbjs.spawn(params).on('error', function (err) {
             console.log('error: ' + err);
         }).on('output', function (output) {
             // console.log('output: ' + output);
         }).on('progress', function (progress) {
-            console.log('ripping: ' + JSON.stringify(progress));
-            //if ((progress) && (progress.taskCount === 2) && (progress.percentComplete !== percentComplete)) {
+            // console.log('ripping: ' + JSON.stringify(progress));
+            // if ((progress) && (progress.taskCount === 2) && (progress.percentComplete !== percentComplete)) {
             //    console.log('ripping to ' + params.output + ' ' + (progress.percentComplete) + '% done');
             //    percentComplete = progress.percentComplete;
-            //}
+            // }
+            if ((progress.taskCount === 2) && (!bar1)) {
+		bar1 = new _cliProgress.Bar({
+			// format : ' |- Ripping ' + inputFileName + ' | {percentage}%' + ' - ' + '||{bar}||',
+			format : 'ripping ' + inputFileName + ' [{bar}] {percentage}% || {value}/{total} Chunks || Speed: {speed}',
+			barCompleteChar: '\u2588',
+			barIncompleteChar: '\u2591',
+			hideCursor : true
+		    });
+
+		    //}, _cliProgress.Presets.shades_classic);
+
+		bar1.start(1000, 0);
+	    }
+            if (parseInt(progress.percentComplete * 10, 10) !== percentComplete) {
+		/*
+		console.log('bar: ');
+		console.log(bar);
+		bar.tick(1);
+		*/
+            	// console.log('ripping to ' + params.output + ' ' + (progress.percentComplete) + '% done');
+
+            	percentComplete = parseInt(progress.percentComplete * 10);
+		bar1.update(percentComplete);
+            }
         }).on('complete', function () {
+            bar1.stop();
             nextRip(null, params);
         });
     }, function (err) {
@@ -236,14 +323,20 @@ getChapters(inputFileLocation, outputFileLocation).then(
         runHandbrake(inputFileLocation, outputFileLocation, chapters);
     }
 );
-// 
-copyFile(dotMD5FileLocation, BACKUP_DIRECTORY + dotMD5FileName);
+//
+/*
+if (fs.existsSync(dotMD5FileLocation)) {
+	copyFile(dotMD5FileLocation, BACKUP_DIRECTORY + dotMD5FileName);
+}
 if (fs.existsSync(BACKUP_DIRECTORY + inputFileName)) {
     console.log('file already exists at: ' + BACKUP_DIRECTORY + inputFileName);
 } else {
     copyFile(inputFileLocation, BACKUP_DIRECTORY + inputFileName);
 }
-copyFile(dotDVDFileLocation, BACKUP_DIRECTORY + dotDVDFileName);
+if (fs.existsSync(dotDVDFileLocation)) {
+	copyFile(dotDVDFileLocation, BACKUP_DIRECTORY + dotDVDFileName);
+}
+*/
 
 
 //watcher.on('add', processISO);
